fix(gates): move the GUI status-judgment baseline into tools/gate-allowlists where exemption data lives
